5 Travel Planner/lab/requirements.txt
google-generativeai>=0.5.0
requests>=2.32.0
python-dotenv>=1.0.1
rich>=13.7.1
tenacity>=8.2.3
openai>=1.0.0
